+++ src/layouts/ProtectedLayout.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from 'react-router-dom';
-import { useAuthStatus } from '../hooks/useAuthStatus.tsx';
+import { useAuthStatus } from '../hooks/auth/useAuthStatus.tsx';
 import { FullscreenLoader } from '../components/common/FullscreenLoader.tsx';
 
 export const ProtectedLayout = () => {
