+++ src/pages/HomePage.tsx
@@ -1,6 +1,6 @@
 import { Link, useNavigate } from 'react-router-dom';
 
-import { useAuthStatus } from '../hooks/useAuthStatus.tsx';
+import { useAuthStatus } from '../hooks/auth/useAuthStatus.tsx';
 import { TripCard } from '../components/trip/TripCard.tsx';
 import { TRIP_IMAGE_PATHS } from '../constants/tripImages.ts';
 import { formatDateRange } from '../utils/common/date.ts';
