+++ src/pages/LoginPage.tsx
@@ -7,7 +7,7 @@ import { useState } from 'react';
 import { useEffect } from 'react';
 import { Navigate } from 'react-router-dom';
 
-import { useAuthStatus } from '../hooks/useAuthStatus.tsx';
+import { useAuthStatus } from '../hooks/auth/useAuthStatus.tsx';
 import { Button } from '../components/common/Button.tsx';
 import type { Provider } from '../types/user.ts';
 import { Typography } from '../components/common/Typography.tsx';
